extracting main module
