+++ pom.xml
@@ -49,6 +49,7 @@
 		<module>modules/filesystem</module>
 		<module>runtime</module>
 		<module>test-support</module>
+		<module>main</module>
 	</modules>
 	<dependencies>
 		<dependency>
