+++ runtime/pom.xml
@@ -17,49 +17,12 @@
 		</dependency>
 		<dependency>
 			<groupId>dfa1</groupId>
-			<artifactId>hosh-filesystem-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
-			<artifactId>hosh-history-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
-			<artifactId>hosh-network-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
-			<artifactId>hosh-system-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
-			<artifactId>hosh-terminal-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
-			<artifactId>hosh-text-module</artifactId>
-			<version>${project.version}</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>dfa1</groupId>
 			<artifactId>hosh-test-support</artifactId>
 			<version>${project.version}</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
 	<build>
-		<finalName>hosh</finalName>
 		<plugins>
 			<plugin>
 				<groupId>org.antlr</groupId>
@@ -74,43 +37,6 @@
 						<goals>
 							<goal>antlr4</goal>
 						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-shade-plugin</artifactId>
-				<version>3.2.2</version>
-				<executions>
-					<execution>
-						<phase>package</phase>
-						<goals>
-							<goal>shade</goal>
-						</goals>
-						<configuration>
-							<createDependencyReducedPom>false</createDependencyReducedPom>
-							<transformers>
-								<transformer implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
-									<resource>LICENSE.md</resource>
-									<file>../LICENSE.md</file>
-								</transformer>
-								<transformer implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
-									<resource>README.md</resource>
-									<file>../README.md</file>
-								</transformer>
-								<transformer implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
-									<resource>CHANGELOG.md</resource>
-									<file>../CHANGELOG.md</file>
-								</transformer>
-								<transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-									<manifestEntries>
-										<Main-Class>hosh.runtime.Hosh</Main-Class>
-										<Add-Exports>java.base/jdk.internal.misc</Add-Exports>
-									</manifestEntries>
-								</transformer>
-								<transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
-							</transformers>
-						</configuration>
 					</execution>
 				</executions>
 			</plugin>
